requirements.txt
discord.py
flask
supabase
python-dotenv